% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_percent.R
\name{missing_percent}
\alias{missing_percent}
\title{Find percentage missing for each variable}
\usage{
missing_percent(
  data_root_or_file,
  output_path = NA,
  file.pattern = ".csv",
  save_report = FALSE
)
}
\arguments{
\item{data_root_or_file}{Directory path to scan for files OR direct file path if single file}

\item{output_path}{Path for saving report. Path must include .csv name of report. Default is same directory + new folder called reports/missing_report.csv}

\item{file.pattern}{Pattern to use for searching files (default: .csv)}
}
\value{
data.frame with three columns: variable, year and missing %
}
\description{
Find percentage missing for each variable
}
