% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surgeon_count.R
\name{surgeon_count}
\alias{surgeon_count}
\title{Count the number of surgeons each year}
\usage{
surgeon_count(data_root, output_path = NA, file.pattern = ".csv")
}
\arguments{
\item{data_root}{Directory path to scan for files}

\item{output_path}{Path for saving report. Path must include .csv name of report. Default is same directory + new folder called reports/surgeon_count_report.csv}

\item{file.pattern}{Pattern to use for searching files (default: .csv)}
}
\value{
data.frame with two columns where each row contains the file and number of observations in the file
}
\description{
Count the number of surgeons each year
}
