% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedure_count.R
\name{procedure_count}
\alias{procedure_count}
\title{Count the number of claims for each procedures each year}
\usage{
procedure_count(
  data_root_or_file,
  output_path = NA,
  file.pattern = ".csv",
  save_report = FALSE
)
}
\arguments{
\item{data_root_or_file}{Directory path to scan for files OR direct file path if single file}

\item{output_path}{Path for saving report. Path must include .csv name of report. Default is same directory + new folder called reports/surgeon_count_report.csv}

\item{file.pattern}{Pattern to use for searching files (default: .csv)}
}
\value{
data.frame with three columns
}
\description{
Count the number of claims for each procedures each year
}
